Changes Read to receive uint64 byte count.

+++ imports/wasi_snapshot_preview1/sock.go
@@ -85,7 +85,7 @@ func sockRecvFn(_ context.Context, mod api.Module, params []uint64) sys.Errno {
 		if !ok {
 			return sys.EINVAL
 		}
-		firstIovecBuf, ok := mem.Read(firstIovecBufAddr, firstIovecBufLen)
+		firstIovecBuf, ok := mem.Read(firstIovecBufAddr, uint64(firstIovecBufLen))
 		if !ok {
 			return sys.EINVAL
 		}
