+++ internal/integration_test/vs/runtime.go
@@ -80,7 +80,7 @@ func (m *wazeroModule) Memory() []byte {
 }
 
 func (r *wazeroRuntime) log(_ context.Context, mod api.Module, stack []uint64) {
-	offset, byteCount := uint32(stack[0]), uint32(stack[1])
+	offset, byteCount := uint32(stack[0]), stack[1]
 
 	buf, ok := mod.Memory().Read(offset, byteCount)
 	if !ok {
